+++ modules/SettingsDialog.js
@@ -6,6 +6,7 @@ define( function( require, exports ) {
 		
 		// Extension Modules.
 		Defaults = require( 'modules/Defaults' ),
+		Strings = require( 'modules/Strings' ),
 		
 		// Templates.
 		settingsDialogTemplate = require( 'text!../html/settings-dialog.html' ),
@@ -65,9 +66,7 @@ define( function( require, exports ) {
 			var $this = $( element );
 			
 			// Make checkbox checked if its value is in array.
-			if ( valueArray.indexOf( $this.attr( 'value' ) ) !== -1 ) {
-				$this.prop( 'checked', true );
-			}
+			$this.prop( 'checked', valueArray.indexOf( $this.attr( 'value' ) ) !== -1 );
 		} );
 	}
 	
@@ -89,7 +88,10 @@ define( function( require, exports ) {
 	 */
 	exports.show = function( prefs ) {
 		// Compile dialog template.
-		var compiledTemplate = Mustache.render( settingsDialogTemplate, {} );
+		var compiledTemplate = Mustache.render( settingsDialogTemplate, {
+			Strings: Strings,
+			phpAvailable: prefs.get( 'php-available' )
+		} );
 		
 		// Save dialog to variable.
 		dialog = Dialogs.showModalDialogUsingTemplate( compiledTemplate );