+++ nls/root/Strings.js
@@ -1,0 +1,15 @@
+define( {
+	// EXTENSION.
+	EXTENSION_NAME: "PHP Code Quality Tools",
+	
+	// GENERIC.
+	OK:     "OK",
+	CANCEL: "Cancel",
+	RESET:  "Reset",
+	
+	// SETTINGS DIALOG
+	SETTINGS_DIALOG_TITLE:         'PHP Code Quality Tools Settings',
+	SETTINGS_DIALOG_ENABLE:        'Enable',
+	SETTINGS_DIALOG_STANDARDS:     'Coding standards',
+	SETTINGS_DIALOG_NOT_AVAILABLE: 'PHP is not available on your machine. The extensions will not be able to run any tools.'
+} );