+++ nls/sv/Strings.js
@@ -1,0 +1,15 @@
+define( {
+	// EXTENSION.
+	EXTENSION_NAME: "PHP Code Quality Tools",
+	
+	// GENERIC.
+	OK:     "OK",
+	CANCEL: "Avbryt",
+	RESET:  "Återställ",
+	
+	// SETTINGS DIALOG
+	SETTINGS_DIALOG_TITLE:         'Inställningar för PHP Code Quality Tools',
+	SETTINGS_DIALOG_ENABLE:        'Aktivera',
+	SETTINGS_DIALOG_STANDARDS:     'Kodstandarder',
+	SETTINGS_DIALOG_NOT_AVAILABLE: 'PHP är inte tillgängligt på din maskin. Tillägget kommer inte kunna köra några verktyg.'
+} );