+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# Brackets PHP Lint Tools Changelog
+# Brackets PHP Code Quality Tools Changelog
+
+## 0.1.2
+* Only collect errors in PHP files.
+
+## 0.1.1
+* Fix issue with loading order of modules on Windows.
 
 ## 0.1.0
 * Initial release.